+++ libp2p/kad_dht/kad_dht.py
@@ -18,11 +18,12 @@ from multiaddr import (
 import trio
 import varint
 
+from libp2p import env_to_send_in_RPC
 from libp2p.abc import (
     IHost,
 )
 from libp2p.discovery.random_walk.rt_refresh_manager import RTRefreshManager
-from libp2p.kad_dht.utils import env_to_send_in_RPC, maybe_consume_signed_record
+from libp2p.kad_dht.utils import maybe_consume_signed_record
 from libp2p.network.stream.net_stream import (
     INetStream,
 )
